requirements.txt
redis
asyncio
aiogram
requests
python-dotenv
